+++ db/migrate/20150618133414_add_deleted_by_id_to_posts.rb
@@ -1,0 +1,6 @@
+class AddDeletedByIdToPosts < ActiveRecord::Migration
+  def change
+    add_column :posts, :deleted_by_id, :integer
+    add_index :posts, :deleted_by_id
+  end
+end